Remove utilities to SomeUtil.jl

+++ src/AbstractXAgent.jl
@@ -14,7 +14,7 @@ export AbstractAgent, AbstractXAgent
 export verify, getIDCOUNTER
 
 "The ID assigned to an agent for every new agent"
-global IDCOUNTER = 0::Int64              # This is differnt than agents.jl 
+global IDCOUNTER = 0::Int              # This is differnt than agents.jl 
 
 "A counter for assigning unique ID for each agent"
 getIDCOUNTER() = global IDCOUNTER = IDCOUNTER + 1
